// src/app/shared/enum/refTable.enum.ts
enum ReferenceTable {
    Facility_Type = 'FACILITY_TYPE',
    Garage_Type = 'SURFACE_TYPE',
    Service_Type = 'SERVICE_TYPE',
    Supported_Vehicle_Type = 'SUPPORTED_VEHICLE_TYPE',
    Available_Amenities = 'AMENITY',
    Location_Signage = 'SIGNAGE',
    Payment_Options = 'PAYMENT_OPTION',
    Contact_Type = 'CONTACT_TYPE',
    Company = 'COMPANY',
    Address_Type = 'ADDRESS_TYPE',
    Validations_Methods = 'VALIDATION_METHOD',
    Note_Type = 'NOTE_TYPE',
    Vertical_Markets_Position = 'VERTICAL_MARKET_POSITION',
    National_Account = 'NATIONAL_ACCOUNT',
    Rate_Type = 'RATE_TYPE',
    Increment = 'INCREMENT',
    Location_Attributes='ATTRIBUTE',
    Facility_Access = 'FACILITY_ACCESS_TYPE',    
    RevenueSystem='REVENUE_APPLICATION',  
    Vertical_Markets='VERTICAL_MARKET'      
};
export default ReferenceTable;